requirements.txt
streamlit
supabase
python-dotenv
requests
moviepy
